requirements.txt
Django==4.0.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
docType==0.23
nltk==3.7
whitenoise==6.2.0
gunicorn==20.1.0
django-ckeditor==6.5.0
beautifulsoup4==4.11.1
lxml==4.9.1